requirements.txt
langgraph
langchain
langchain-anthropic
langchain-community
tavily-python
python-dotenv

# Optional visualization dependencies
# matplotlib
# graphviz
